Swap out LiteDB implementation for SQLite
