--- src/SystemStateManager.Persistence.SQLite/IPersistentSnapshot.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace DevOptimal.SystemStateManager.Persistence.SQLite
-{
-    internal interface IPersistentSnapshot : ISnapshot
-    {
-        int ProcessID { get; }
-
-        DateTime ProcessStartTime { get; }
-    }
-}
