--- src/SystemStateManager.Persistence.SQLite/SystemStateManager.Persistence.SQLite.csproj
@@ -1,23 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <Description>Library for snapshotting and restoring the state of system resources in a persistent manner.</Description>
-    <PackageTags>$(PackageTags);persistence</PackageTags>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
-      <_Parameter1>DevOptimal.SystemStateManager.Persistence.Tests, PublicKey=002400000480000094000000060200000024000052534131000400000100010051d5ac740ee660ab5bd98dd1002f26432af95c197165d891388550c70a79bd35907820cbe6a43b652876304cdc3e847e4543ad43d5289dd01a8377ab946be9a8f48be16b6136a1ea33d03d1d4b4767746d51536876249ce28673c826c516bd8965b8518826377e58407716919895b4b962ca8d810cab747ca0212b94da6b89df</_Parameter1>
-    </AssemblyAttribute>
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Data.Sqlite" Version="7.0.3" />
-    <PackageReference Include="System.IO.FileSystem.AccessControl" Version="5.0.0" Condition="'$(TargetFramework)' == 'netstandard2.0'" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\SystemStateManager\SystemStateManager.csproj" />
-  </ItemGroup>
-
-</Project>
