--- src/SystemStateManager.Persistence.SQLite/version.json
@@ -1,8 +1,0 @@
-{
-    "$schema": "https://raw.githubusercontent.com/dotnet/Nerdbank.GitVersioning/main/src/NerdBank.GitVersioning/version.schema.json",
-    "inherit": true,
-    "pathFilters": [
-        ".",
-        "../SystemStateManager"
-    ]
-}
