+++ src/SystemStateManager.Persistence/FileSystem/Caching/SQLiteFileCache.cs
