+++ src/SystemStateManager.Persistence/FileSystem/PersistentFileCaretaker.cs
@@ -1,52 +1,102 @@
 ﻿using DevOptimal.SystemStateManager.FileSystem;
-using LiteDB;
+using DevOptimal.SystemUtilities.FileSystem;
+using Microsoft.Data.Sqlite;
 using System;
-using System.Linq;
+using System.Collections.Generic;
 
-namespace DevOptimal.SystemStateManager.Persistence.FileSystem
+namespace DevOptimal.SystemStateManager.Persistence.SQLite.FileSystem
 {
-    internal class PersistentFileCaretaker : PersistentCaretaker<PersistentFileOriginator, FileMemento>
+    internal class PersistentFileCaretaker : PersistentCaretaker<FileOriginator, FileMemento>
     {
-        private bool disposedValue;
-
-        public PersistentFileCaretaker(string id, PersistentFileOriginator originator)
-            : base(id, originator)
+        public PersistentFileCaretaker(string id, FileOriginator originator, SqliteConnection connection)
+            : base(id, originator, connection)
         {
         }
 
-        [BsonCtor]
-        public PersistentFileCaretaker(string _id, int processID, DateTime processStartTime, BsonDocument originator, BsonDocument memento)
-            : base(_id, processID, processStartTime, LiteDatabaseFactory.Mapper.ToObject<PersistentFileOriginator>(originator), LiteDatabaseFactory.Mapper.ToObject<FileMemento>(memento))
+        public PersistentFileCaretaker(string id, int processID, DateTime processStartTime, FileOriginator originator, FileMemento memento, SqliteConnection connection)
+            : base(id, processID, processStartTime, originator, memento, connection)
         {
         }
 
-        protected override void Dispose(bool disposing)
-        {
-            base.Dispose(disposing);
 
-            if (!disposedValue)
+        protected override void Initialize()
+        {
+            var command = connection.CreateCommand();
+            command.CommandText =
+            $@"CREATE TABLE IF NOT EXISTS {nameof(PersistentFileCaretaker)} (
+                '{nameof(ID)}' TEXT PRIMARY KEY,
+                '{nameof(ProcessID)}' INTEGER NOT NULL,
+                '{nameof(ProcessStartTime)}' INTEGER NOT NULL,
+                '{nameof(Originator.Path)}' TEXT NOT NULL,
+                '{nameof(Memento.Hash)}' TEXT
+            );";
+            command.ExecuteNonQuery();
+        }
+
+        protected override void Persist()
+        {
+            var command = connection.CreateCommand();
+            command.CommandText =
+            $@"INSERT INTO {nameof(PersistentFileCaretaker)} (
+                '{nameof(ID)}',
+                '{nameof(ProcessID)}',
+                '{nameof(ProcessStartTime)}',
+                '{nameof(Originator.Path)}',
+                '{nameof(Memento.Hash)}'
+            ) VALUES (
+                @{nameof(ID)},
+                @{nameof(ProcessID)},
+                @{nameof(ProcessStartTime)},
+                @{nameof(Originator.Path)},
+                @{nameof(Memento.Hash)}
+            );";
+            command.Parameters.AddWithValue($"@{nameof(ID)}", ID);
+            command.Parameters.AddWithValue($"@{nameof(ProcessID)}", ProcessID);
+            command.Parameters.AddWithValue($"@{nameof(ProcessStartTime)}", ProcessStartTime.Ticks);
+            command.Parameters.AddWithValue($"@{nameof(Originator.Path)}", Originator.Path);
+            command.Parameters.AddWithNullableValue($"@{nameof(Memento.Hash)}", Memento.Hash);
+            command.ExecuteNonQuery();
+        }
+
+        protected override void Unpersist()
+        {
+            var command = connection.CreateCommand();
+            command.CommandText = $@"DELETE FROM {nameof(PersistentFileCaretaker)} WHERE {nameof(ID)} = @{nameof(ID)};";
+            command.Parameters.AddWithValue($"@{nameof(ID)}", ID);
+            command.ExecuteNonQuery();
+        }
+
+        public static IEnumerable<IPersistentSnapshot> GetCaretakers(SqliteConnection connection, IFileSystem fileSystem, IFileCache fileCache)
+        {
+            var caretakers = new List<PersistentFileCaretaker>();
+
+            if (connection.TableExists(nameof(PersistentFileCaretaker)))
             {
-                if (disposing)
+                using (var reader = connection.ExecuteReader($@"SELECT * FROM {nameof(PersistentFileCaretaker)};"))
                 {
-                    if (Memento.Hash != null)
+                    while (reader.Read())
                     {
-                        using (var database = LiteDatabaseFactory.GetDatabase())
-                        {
-                            var collection = database.GetCollection<IPersistentSnapshot>();
-
-                            if (!collection.FindAll().OfType<PersistentFileCaretaker>().Any(c => c.Memento.Hash == Memento.Hash))
+                        var caretaker = new PersistentFileCaretaker(
+                            id: reader.GetString(nameof(ID)),
+                            processID: reader.GetInt32(nameof(ProcessID)),
+                            processStartTime: new DateTime(reader.GetInt64(nameof(ProcessStartTime))),
+                            originator: new FileOriginator(
+                                path: reader.GetString(nameof(FileOriginator.Path)),
+                                fileCache: fileCache,
+                                fileSystem: fileSystem
+                            ),
+                            memento: new FileMemento
                             {
-                                var fileStorage = database.FileStorage;
-                                fileStorage.Delete(Memento.Hash);
-                            }
-                        }
+                                Hash = reader.GetNullableString(nameof(FileMemento.Hash))
+                            },
+                            connection: connection
+                        );
+                        caretakers.Add(caretaker);
                     }
                 }
-
-                // TODO: free unmanaged resources (unmanaged objects) and override finalizer
-                // TODO: set large fields to null
-                disposedValue = true;
             }
+
+            return caretakers;
         }
     }
 }
