--- src/SystemStateManager.Persistence/FileSystem/PersistentFileOriginator.cs
@@ -1,20 +1,0 @@
-﻿using DevOptimal.SystemStateManager.FileSystem;
-using DevOptimal.SystemUtilities.FileSystem;
-using LiteDB;
-
-namespace DevOptimal.SystemStateManager.Persistence.FileSystem
-{
-    internal class PersistentFileOriginator : FileOriginator
-    {
-        public PersistentFileOriginator(string path, IFileCache fileCache, IFileSystem fileSystem)
-            : base(path, fileCache, fileSystem)
-        {
-        }
-
-        [BsonCtor]
-        public PersistentFileOriginator(string path, BsonDocument fileCache, BsonDocument fileSystem)
-            : this(path, LiteDatabaseFactory.Mapper.ToObject<IFileCache>(fileCache), LiteDatabaseFactory.Mapper.ToObject<IFileSystem>(fileSystem))
-        {
-        }
-    }
-}
