+++ src/SystemStateManager.Persistence/IPersistentSnapshot.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace DevOptimal.SystemStateManager.Persistence
+namespace DevOptimal.SystemStateManager.Persistence.SQLite
 {
     internal interface IPersistentSnapshot : ISnapshot
     {
