+++ src/SystemStateManager.Persistence/PersistentSystemStateManager.cs
@@ -1,11 +1,14 @@
-﻿using DevOptimal.SystemStateManager.FileSystem;
-using DevOptimal.SystemStateManager.Persistence.Environment;
-using DevOptimal.SystemStateManager.Persistence.FileSystem;
-using DevOptimal.SystemStateManager.Persistence.FileSystem.Caching;
-using DevOptimal.SystemStateManager.Persistence.Registry;
+﻿using DevOptimal.SystemStateManager.Environment;
+using DevOptimal.SystemStateManager.FileSystem;
+using DevOptimal.SystemStateManager.Persistence.SQLite.Environment;
+using DevOptimal.SystemStateManager.Persistence.SQLite.FileSystem;
+using DevOptimal.SystemStateManager.Persistence.SQLite.FileSystem.Caching;
+using DevOptimal.SystemStateManager.Persistence.SQLite.Registry;
+using DevOptimal.SystemStateManager.Registry;
 using DevOptimal.SystemUtilities.Environment;
 using DevOptimal.SystemUtilities.FileSystem;
 using DevOptimal.SystemUtilities.Registry;
+using Microsoft.Data.Sqlite;
 using Microsoft.Win32;
 using System;
 using System.Collections.Generic;
@@ -13,8 +16,11 @@ using System.ComponentModel;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using System.Runtime.InteropServices;
+using System.Security.AccessControl;
+using System.Security.Principal;
 
-namespace DevOptimal.SystemStateManager.Persistence
+namespace DevOptimal.SystemStateManager.Persistence.SQLite
 {
     public class PersistentSystemStateManager : SystemStateManager
     {
@@ -22,7 +28,9 @@ namespace DevOptimal.SystemStateManager.Persistence
             Path.Combine(
                 System.Environment.GetFolderPath(System.Environment.SpecialFolder.CommonApplicationData),
                 nameof(SystemStateManager),
-                $"{nameof(Persistence)}.litedb"));
+                $"{nameof(Persistence)}.db"));
+
+        private readonly SqliteConnection connection;
 
         public PersistentSystemStateManager()
             : this(new DefaultEnvironment(), new DefaultFileSystem(), new DefaultRegistry())
@@ -30,49 +38,64 @@ namespace DevOptimal.SystemStateManager.Persistence
         }
 
         public PersistentSystemStateManager(IEnvironment environment, IFileSystem fileSystem, IRegistry registry)
-            : base(new LiteDBFileCache(fileSystem), environment, fileSystem, registry)
+            : this(environment, fileSystem, registry, CreateConnection())
         {
         }
 
-        internal PersistentSystemStateManager(List<ISnapshot> snapshots)
-            : base(snapshots, new LiteDBFileCache(new DefaultFileSystem()))
+        private PersistentSystemStateManager(IEnvironment environment, IFileSystem fileSystem, IRegistry registry, SqliteConnection connection)
+            : base(new SQLiteFileCache(fileSystem, connection), environment, fileSystem, registry)
         {
+            this.connection = connection;
+        }
+
+        private PersistentSystemStateManager(List<ISnapshot> snapshots, IFileCache fileCache, IEnvironment environment, IFileSystem fileSystem, IRegistry registry, SqliteConnection connection)
+            : base(snapshots, fileCache, environment, fileSystem, registry)
+        {
+            this.connection = connection;
         }
 
         protected override ISnapshot CreateEnvironmentVariableSnapshot(string id, string name, EnvironmentVariableTarget target, IEnvironment environment)
         {
-            var originator = new PersistentEnvironmentVariableOriginator(name, target, environment);
-            return new PersistentEnvironmentVariableCaretaker(id, originator);
+            var originator = new EnvironmentVariableOriginator(name, target, environment);
+            return new PersistentEnvironmentVariableCaretaker(id, originator, connection);
         }
 
         protected override ISnapshot CreateDirectorySnapshot(string id, string path, IFileSystem fileSystem)
         {
-            var originator = new PersistentDirectoryOriginator(path, fileSystem);
-            return new PersistentDirectoryCaretaker(id, originator);
+            var originator = new DirectoryOriginator(path, fileSystem);
+            return new PersistentDirectoryCaretaker(id, originator, connection);
         }
 
         protected override ISnapshot CreateFileSnapshot(string id, string path, IFileCache fileCache, IFileSystem fileSystem)
         {
-            var originator = new PersistentFileOriginator(path, fileCache, fileSystem);
-            return new PersistentFileCaretaker(id, originator);
+            var originator = new FileOriginator(path, fileCache, fileSystem);
+            return new PersistentFileCaretaker(id, originator, connection);
         }
 
         protected override ISnapshot CreateRegistryKeySnapshot(string id, RegistryHive hive, RegistryView view, string subKey, IRegistry registry)
         {
-            var originator = new PersistentRegistryKeyOriginator(hive, view, subKey, registry);
-            return new PersistentRegistryKeyCaretaker(id, originator);
+            var originator = new RegistryKeyOriginator(hive, view, subKey, registry);
+            return new PersistentRegistryKeyCaretaker(id, originator, connection);
         }
 
         protected override ISnapshot CreateRegistryValueSnapshot(string id, RegistryHive hive, RegistryView view, string subKey, string name, IRegistry registry)
         {
-            var originator = new PersistentRegistryValueOriginator(hive, view, subKey, name, registry);
-            return new PersistentRegistryValueCaretaker(id, originator);
+            var originator = new RegistryValueOriginator(hive, view, subKey, name, registry);
+            return new PersistentRegistryValueCaretaker(id, originator, connection);
         }
 
         /// <summary>
         /// Restores all abandoned snapshots on the current machine. An "abandoned snapshot" is a snapshot that was created by a process that no longer exists.
         /// </summary>
-        public static void RestoreAbandonedSnapshots()
+        public static void RestoreAbandonedSnapshots() => RestoreAbandonedSnapshots(new DefaultEnvironment(), new DefaultFileSystem(), new DefaultRegistry());
+
+        /// <summary>
+        /// Restores all abandoned snapshots on the current machine. An "abandoned snapshot" is a snapshot that was created by a process that no longer exists.
+        /// </summary>
+        /// <param name="environment">A concrete implementation of the machine's environment.</param>
+        /// <param name="fileSystem">A concrete implementation of the machine's file system.</param>
+        /// <param name="registry">A concrete implementation of the machine's registry.</param>
+        public static void RestoreAbandonedSnapshots(IEnvironment environment, IFileSystem fileSystem, IRegistry registry)
         {
             // Create a dictionary that maps process IDs to process start times, which will be used to uniquely identify a currently running process.
             // A null value indicates that the current process does not have permission to the corresponding process - try rerunning in an elevated process.
@@ -90,24 +113,69 @@ namespace DevOptimal.SystemStateManager.Persistence
                 catch (InvalidOperationException) { } // The process has already exited, so don't add it.
             }
 
-            var abandonedSnapshots = new List<ISnapshot>();
+            var connection = CreateConnection();
+            var fileCache = new SQLiteFileCache(fileSystem, connection);
 
-            using (var database = LiteDatabaseFactory.GetDatabase())
-            {
-                abandonedSnapshots.AddRange(database.GetCollection<IPersistentSnapshot>().FindAll()
-                    .Where(c => !(processes.ContainsKey(c.ProcessID) &&
-                        (
-                            processes[c.ProcessID] == c.ProcessStartTime ||
-                            processes[c.ProcessID] == null
-                        )))
-                    .Cast<ISnapshot>());
-            }
+            var allSnapshots = PersistentEnvironmentVariableCaretaker.GetCaretakers(connection, environment)
+                .Concat(PersistentDirectoryCaretaker.GetCaretakers(connection, fileSystem))
+                .Concat(PersistentFileCaretaker.GetCaretakers(connection, fileSystem, fileCache))
+                .Concat(PersistentRegistryKeyCaretaker.GetCaretakers(connection, registry))
+                .Concat(PersistentRegistryValueCaretaker.GetCaretakers(connection, registry))
+                .Where(c => !(processes.ContainsKey(c.ProcessID) && (processes[c.ProcessID] == c.ProcessStartTime || processes[c.ProcessID] == null)));
+
+            var abandonedSnapshots = new List<ISnapshot>(allSnapshots);
 
             if (abandonedSnapshots.Any())
             {
-                var systemStateManager = new PersistentSystemStateManager(abandonedSnapshots);
+                var systemStateManager = new PersistentSystemStateManager(abandonedSnapshots, fileCache, environment, fileSystem, registry, connection);
                 systemStateManager.Dispose();
             }
+        }
+
+        private static SqliteConnection CreateConnection()
+        {
+            if (!PersistenceURI.IsFile)
+            {
+                throw new NotSupportedException($"{nameof(PersistenceURI)} is invalid. Only local file paths are supported.");
+            }
+
+            var databaseFile = new FileInfo(PersistenceURI.LocalPath);
+
+            var databaseDirectory = databaseFile.Directory;
+
+            if (!databaseDirectory.Exists)
+            {
+                databaseDirectory.Create();
+
+                if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+                {
+                    var directorySecurity = databaseDirectory.GetAccessControl();
+                    directorySecurity.AddAccessRule(new FileSystemAccessRule(
+                        identity: new SecurityIdentifier(WellKnownSidType.WorldSid, domainSid: null),
+                        fileSystemRights: FileSystemRights.FullControl,
+                        inheritanceFlags: InheritanceFlags.ContainerInherit | InheritanceFlags.ObjectInherit,
+                        propagationFlags: PropagationFlags.NoPropagateInherit,
+                        type: AccessControlType.Allow));
+                    databaseDirectory.SetAccessControl(directorySecurity);
+                }
+            }
+
+            var connectionString = new SqliteConnectionStringBuilder
+            {
+                DataSource = databaseFile.FullName,
+                Cache = SqliteCacheMode.Shared,
+                Mode = SqliteOpenMode.ReadWriteCreate
+            }.ToString();
+            var connection = new SqliteConnection(connectionString);
+            connection.Open();
+
+            return connection;
+        }
+
+        protected override void Dispose(bool disposing)
+        {
+            base.Dispose(disposing);
+            connection.Dispose();
         }
     }
 }
