+++ src/SystemStateManager.Persistence/ResourceLockedException.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace DevOptimal.SystemStateManager.Persistence
+namespace DevOptimal.SystemStateManager.Persistence.SQLite
 {
     public class ResourceLockedException : Exception
     {
