+++ src/SystemStateManager/SystemStateManager.csproj
@@ -8,9 +8,6 @@
     <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
       <_Parameter1>DevOptimal.SystemStateManager.Persistence, PublicKey=002400000480000094000000060200000024000052534131000400000100010051d5ac740ee660ab5bd98dd1002f26432af95c197165d891388550c70a79bd35907820cbe6a43b652876304cdc3e847e4543ad43d5289dd01a8377ab946be9a8f48be16b6136a1ea33d03d1d4b4767746d51536876249ce28673c826c516bd8965b8518826377e58407716919895b4b962ca8d810cab747ca0212b94da6b89df</_Parameter1>
     </AssemblyAttribute>
-    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
-      <_Parameter1>DevOptimal.SystemStateManager.Persistence.SQLite, PublicKey=002400000480000094000000060200000024000052534131000400000100010051d5ac740ee660ab5bd98dd1002f26432af95c197165d891388550c70a79bd35907820cbe6a43b652876304cdc3e847e4543ad43d5289dd01a8377ab946be9a8f48be16b6136a1ea33d03d1d4b4767746d51536876249ce28673c826c516bd8965b8518826377e58407716919895b4b962ca8d810cab747ca0212b94da6b89df</_Parameter1>
-    </AssemblyAttribute>
   </ItemGroup>
 
   <ItemGroup>
