--- test/SystemStateManager.Persistence.SQLite.Tests/PrimaryKeyAttribute.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace DevOptimal.SystemStateManager.Persistence.SQLite.Tests
-{
-    [AttributeUsage(AttributeTargets.Property, AllowMultiple = false, Inherited = false)]
-    public class PrimaryKeyAttribute : Attribute
-    {
-    }
-}
