--- test/SystemStateManager.Persistence.SQLite.Tests/SystemStateManager.Persistence.SQLite.Tests.csproj
@@ -1,6 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Data.Sqlite" Version="7.0.3" />
-  </ItemGroup>
-</Project>
